refactor(world): split world_map into WorldMapController + WorldMapView + signals (audit B-1)
