+++ lib/features/quests/user_stars.dart
@@ -10,8 +10,9 @@ import 'package:fluffychat/routes/chat/choreographer/activity_orchestrator/orche
 ///
 /// The single source for progression-gate stars: the activity start page
 /// (`not_started_session_controller`) and the world map
-/// (`world_map._deriveActivitySignals`) both feed the gate from this function, so
-/// their lock state cannot drift (quests.instructions.md). The top-right cluster's
+/// (`world_map_signals.deriveActivitySignals`) both feed the gate from this
+/// function, so their lock state cannot drift (quests.instructions.md). The
+/// top-right cluster's
 /// running total (`world_user_cluster._totalStars`) is computed separately, from
 /// `orchestratorAwardedGoals`, because it must work before the activity plan
 /// hydrates (this function needs `ownRole`, which needs the plan); keep that one
